fix ID generation big change

+++ src/main/java/com/cricket/fantasyleague/payload/response/UserTeamRequest.java
@@ -1,4 +1,4 @@
 package com.cricket.fantasyleague.payload.response;
 
-public record UserTeamRequest(Integer userId, Integer matchId) {
+public record UserTeamRequest(Long userId, Integer matchId) {
 }
