+++ src/main/java/com/cricket/fantasyleague/repository/UserMatchStatsDraftRespository.java
@@ -11,17 +11,17 @@ import com.cricket.fantasyleague.entity.table.Match;
 import com.cricket.fantasyleague.entity.table.User;
 import com.cricket.fantasyleague.entity.table.UserMatchStatsDraft;
 
-public interface UserMatchStatsDraftRespository extends JpaRepository<UserMatchStatsDraft,Integer> 
+public interface UserMatchStatsDraftRespository extends JpaRepository<UserMatchStatsDraft, Long> 
 {
     UserMatchStatsDraft findByMatchidAndUserid(Match matchid, User userid);
     List<UserMatchStatsDraft> findAllByMatchid(Match matchid);
 
     @Query("SELECT d.id FROM UserMatchStatsDraft d WHERE d.matchid = :match ORDER BY d.id")
-    List<Integer> findIdsByMatchid(@Param("match") Match match, Pageable pageable);
+    List<Long> findIdsByMatchid(@Param("match") Match match, Pageable pageable);
 
     @Query("SELECT d FROM UserMatchStatsDraft d LEFT JOIN FETCH d.playing11 " +
            "WHERE d.id IN :ids ORDER BY d.id")
-    List<UserMatchStatsDraft> findAllByIdInWithPlaying11(@Param("ids") List<Integer> ids);
+    List<UserMatchStatsDraft> findAllByIdInWithPlaying11(@Param("ids") List<Long> ids);
 
     long countByMatchid(Match match);
 }
